mmsegmentation/requirements.txt
addict==2.4.0
brotlipy==0.7.0
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.6
cmake==3.27.0
colorama==0.4.6
contourpy==1.1.0
cryptography @ file:///croot/cryptography_1689373676338/work
cycler==0.11.0
filelock @ file:///croot/filelock_1672387128942/work
fonttools==4.41.1
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.8.0
importlib-resources==6.0.0
Jinja2 @ file:///croot/jinja2_1666908132255/work
kiwisolver==1.4.4
lit==16.0.6
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.7.2
mdurl==0.1.2
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
mmcv==2.0.1
mmcv-full==1.7.1
mmengine==0.8.2
mmsegmentation==1.1.1
model-index==0.1.11
mpmath==1.2.1
networkx @ file:///croot/networkx_1678964333703/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.74
opendatalab==0.0.9
openmim==0.3.9
ordered-set==4.1.0
packaging==23.1
pandas==2.0.3
Pillow==9.4.0
platformdirs==3.9.1
prettytable==3.8.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.18.0
Pygments==2.15.1
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests @ file:///croot/requests_1690400202158/work
rich==13.4.2
scipy==1.10.1
six==1.16.0
sympy @ file:///croot/sympy_1668202399572/work
tabulate==0.9.0
termcolor==2.3.0
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
urllib3 @ file:///croot/urllib3_1686163155763/work
wcwidth==0.2.6
yapf==0.40.1
zipp==3.16.2
